Redesign custom group mapping
The original custom group mapping was designed for an O(1) lookup of the custom group's URI using the custom group's name as a key. With REXX, special consideration is needed for lowercase names as lowercase names will be converted to uppercase when appending the custom group's name as a key or stem tail value. Therefore, a redesign of the custom user group's mapping was required reducing the performance of the lookup time to O(N)

Signed-off-by: Dominic Schroeder <Dominic.Schroeder@ibm.com>